+++ ExtendableEnums.TestHost/Controllers/OData/SampleBooksController.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text.Json;
 using ExtendableEnums.Testing.Models;
-using Microsoft.AspNet.OData;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.AspNetCore.OData.Query;
+using Microsoft.AspNetCore.OData.Routing.Controllers;
 using Newtonsoft.Json;
 
 namespace ExtendableEnums.TestHost.Controllers.OData;
@@ -13,12 +14,27 @@ public class SampleBooksController : ODataController
 {
     private readonly IList<SampleBook> books = DataContext.Books;
 
+    [HttpGet]
     [EnableQuery]
-    public IEnumerable<SampleBook> Get()
+    public IActionResult Get()
     {
-        return books;
+        return Ok(books);
     }
 
+    [HttpGet("odata/SampleBooks({id})")]
+    [EnableQuery]
+    public IActionResult Get(string id)
+    {
+        var book = books.FirstOrDefault(x => x.Id == id);
+        if (book is null)
+        {
+            return NotFound($"Cannot find book with Id='{id}'.");
+        }
+
+        return Ok(book);
+    }
+
+    [HttpPost]
     [EnableQuery]
     public IActionResult Post([FromBody] JsonElement json)
     {
@@ -28,6 +44,11 @@ public class SampleBooksController : ODataController
         }
 
         var book = JsonConvert.DeserializeObject<SampleBook>(json.GetRawText());
+
+        if (book is null)
+        {
+            throw new ArgumentException("Unable to deserialize the json parameter.", nameof(json));
+        }
 
         var matchingBook = books.FirstOrDefault(b => b.Id == book.Id);
         if (matchingBook is null)