+++ ExtendableEnums.Testing.Models/SampleBook.cs
@@ -3,14 +3,8 @@
 namespace ExtendableEnums.Testing.Models;
 
 [DataContract]
-public class SampleBook
+public class SampleBook : SampleBookBase
 {
-    [DataMember(Name = "id")]
-    public string? Id { get; set; }
-
     [DataMember(Name = "status")]
     public SampleStatus? Status { get; set; }
-
-    [DataMember(Name = "title")]
-    public string? Title { get; set; }
 }